requirements.txt
cycler==0.10.0
getch==1.0
install==1.3.4
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
pyfiglet==0.8.post1
pyparsing==2.4.7
python-dateutil==2.8.1
selenium==3.141.0
six==1.15.0
torch==1.8.0+cpu
torchaudio==0.8.0
torchvision==0.9.0+cpu
typing-extensions==3.7.4.3
urllib3==1.26.4
